Turn asserts in ASIdentifierChoice_canonize() into error checks
The first assert ensures that a stack that was just sorted in a stronger
sense is sorted in a weak sense and the second assert ensures that
the result of the canonization procedure is canonical. All callers check
for error, so these asserts don't do anything useful.

ok jsing

(cherry picked from commit 920304fa6a24b5094b74afa29d3e6eefa753fb55)