handle the status ring entries as 64bit words instead of a struct.
the status ring entries are 8 bytes/64bit, and depending on the
type of entry it has fields all over the place. this loads the
descriptor with a single 64bit read, and then shifts and masks the
bits out of it depending on the type of descriptor. this looks
cleaner for the tx completions in particular.

(cherry picked from commit f86fe765762b07cae5866f7a06815314f09e9feb)