"Unregister" installed firmware that isn't in the SHA256.sig
We assume in this case that the firmware's license was improved and
it moved to be distributed in the base system.  If we find that situation
remove the package registration but leave the firmware files.

(cherry picked from commit 1e63d4ee84c2a7099f543f4e61b984e09c2fc9ab)